% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{df_master}
\alias{df_master}
\title{Master data frame for all object to cache}
\value{
master sheet of sesameData objects
}
\description{
This is an internal object which will be updated on every new release
library(ExperimentHub)
eh <- query(ExperimentHub(localHub=FALSE), c("sesameData", "v1.13.1"))
data.frame(name=eh$title, eh=names(eh))
}
\details{
Cache location is default to
/Users/zhouw3/Library/Caches/org.R-project.R/R/ExperimentHub/
}
