% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend}
\alias{extend}
\title{Extend a GRanges}
\usage{
extend(gr, upstream = 0, downstream = 0)
}
\arguments{
\item{gr}{a GenomicRanges::GRanges}

\item{upstream}{distance to expand upstream}

\item{downstream}{distance to expand downstream}
}
\value{
a GenomicRanges::GRanges
}
\description{
source: https://support.bioconductor.org/p/78652/
}
