% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{inferPlatformFromProbeIDs}
\alias{inferPlatformFromProbeIDs}
\title{infer platform from Probe_IDs}
\usage{
inferPlatformFromProbeIDs(Probe_IDs, silent = FALSE)
}
\arguments{
\item{Probe_IDs}{probe IDs}

\item{silent}{suppress message}
}
\value{
a platform code
}
\description{
infer platform from Probe_IDs
}
\examples{
sesameDataCache("probeIDSignature")
inferPlatformFromProbeIDs(c("cg14620903","cg22464003"))
}
