% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{sesameData_check_platform}
\alias{sesameData_check_platform}
\title{Check platform code}
\usage{
sesameData_check_platform(platform = NULL, probes = NULL)
}
\arguments{
\item{platform}{input platform}

\item{probes}{probes by which the platform may be guessed}
}
\value{
platform code
}
\description{
Note: custome platforms lead to error here.
}
\examples{
sesameData_check_platform("HM450")
}
