% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{sesameData_getGenomeInfo}
\alias{sesameData_getGenomeInfo}
\title{Get genome info files}
\usage{
sesameData_getGenomeInfo(genome)
}
\arguments{
\item{genome}{hg38, mm10, or GRanges with a metadata(genome)[["genome"]]}
}
\value{
a list of genome info files
}
\description{
Get genome info files
}
\examples{
sesameDataCache("genomeInfo.hg38")
res <- sesameData_getGenomeInfo("hg38")
}
