% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameData_getManifestGRanges.R
\name{sesameData_getManifestGRanges}
\alias{sesameData_getManifestGRanges}
\title{get Infinium manifest GRanges}
\usage{
sesameData_getManifestGRanges(platform, genome = NULL)
}
\arguments{
\item{platform}{Mammal40, MM285, EPIC, and HM450}

\item{genome}{hg38, mm10, ... will infer if not given.
For additional mapping, download the GRanges object from
http://zwdzwd.github.io/InfiniumAnnotation
and provide the following argument
..., genome = sesameAnno_buildManifestGRanges("downloaded_file"),...
to this function.}
}
\value{
GRanges
}
\description{
Note that some unaligned probes are not included.
For full manifest, please visit
\url{http://zwdzwd.github.io/InfiniumAnnotation}
}
\examples{
sesameDataCache("Mammal40.address")
res <- sesameData_getManifestGRanges("Mammal40")
}
