% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{sesameData_txnToGeneGRanges}
\alias{sesameData_txnToGeneGRanges}
\title{convert transcript GRanges to gene GRanges}
\usage{
sesameData_txnToGeneGRanges(txns)
}
\arguments{
\item{txns}{GRanges object}
}
\value{
a GRanges object
}
\description{
convert transcript GRanges to gene GRanges
}
\examples{
txns <- sesameData_getTxnGRanges("mm10")
genes <- sesameData_txnToGeneGRanges(txns)

}
