% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataCache.R
\name{sesameDataCache}
\alias{sesameDataCache}
\title{Cache SeSAMe data}
\usage{
sesameDataCache(data_titles = NULL)
}
\arguments{
\item{data_titles}{data to cache, if not given will cache all}
}
\value{
TRUE
}
\description{
Cache SeSAMe data
}
\examples{
sesameDataCache("genomeInfo.hg38")
}
