% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataCache.R
\name{sesameDataCacheAll}
\alias{sesameDataCacheAll}
\title{Cache all SeSAMe data}
\usage{
sesameDataCacheAll(data_titles = NULL)
}
\arguments{
\item{data_titles}{data to cache, if not given will cache all}
}
\value{
TRUE
}
\description{
Cache all SeSAMe data
}
\examples{
sesameDataCache("genomeInfo.hg38")
}
