% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataGet.R
\name{sesameDataGet}
\alias{sesameDataGet}
\title{Get SeSAMe data}
\usage{
sesameDataGet(title, verbose = FALSE)
}
\arguments{
\item{title}{title of the data}

\item{verbose}{whether to output ExperimentHub message}
}
\value{
data object
}
\description{
Get SeSAMe data
}
\examples{

sesameDataCache("EPIC.1.SigDF")
EPIC.1.SigDF <- sesameDataGet('EPIC.1.SigDF')
}
