% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{sesameDataGet_checkEnv}
\alias{sesameDataGet_checkEnv}
\title{Check whether the title exists in cacheEnv}
\usage{
sesameDataGet_checkEnv(title)
}
\arguments{
\item{title}{the title to check}
}
\value{
the data associated with the title or NULL if title doesn't exist
}
\description{
Check whether the title exists in cacheEnv
}
