% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{sesameDataGet_resetEnv}
\alias{sesameDataGet_resetEnv}
\title{Empty cache environment to free memory}
\usage{
sesameDataGet_resetEnv()
}
\value{
gc() output
}
\description{
When this function is called sesameDataGet will
retrieve all data from disk again instead of using the in-memory
cache, i.e., sesameData:::cacheEnv.
}
\details{
Note this is different from sesameDataClearHub which empties the
ExperimentHub on disk.
}
\examples{
sesameDataGet_resetEnv()
}
