% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataGet.R
\name{sesameDataHas}
\alias{sesameDataHas}
\title{Whether sesameData has}
\usage{
sesameDataHas(data_titles)
}
\arguments{
\item{data_titles}{data titles to check}
}
\value{
a boolean vector the same length as data_titles
}
\description{
Whether sesameData has
}
\examples{
sesameDataHas(c("EPIC.address","EPIC.address.Nonexist"))
}
